Factor out common code from two datasources.
