+++ myreports/tests/test_contacts_datasource.py
@@ -1,11 +1,7 @@
 from datetime import datetime
 
-from unittest import TestCase
-
 from myreports.datasources.contacts import (
-    ContactsDataSource, ContactsFilter, ContactsDataSourceJsonDriver)
-from myreports.report_configuration import (
-    ReportConfiguration, ColumnConfiguration)
+    ContactsDataSource, ContactsFilter)
 
 from myjobs.tests.setup import MyJobsBase
 from myjobs.tests.factories import UserFactory
@@ -381,132 +377,3 @@ class MockDataSource(object):
     def run(self, company, filter, order):
         self.calls.append(['run', company, filter, order])
         return 'aaa'
-
-
-class TestContactsDataSourceJsonDriver(TestCase):
-    def setUp(self):
-        super(TestContactsDataSourceJsonDriver, self).setUp()
-        self.ds = MockDataSource()
-        self.driver = ContactsDataSourceJsonDriver(self.ds)
-
-    def test_help_cities(self):
-        """Test that city help gets plumbed to datasource correctly."""
-        result = self.driver.help('company', '{}', 'city', 'zzz')
-        self.assertEqual('aaa', result)
-        self.assertEqual([
-            ['help_city', 'company', ContactsFilter(), 'zzz']
-            ], self.ds.calls)
-
-    def test_run(self):
-        """Test that json run calls are plumbed to datasource correctly."""
-        result = self.driver.run('company', '{}', '["zzz"]')
-        self.assertEqual('aaa', result)
-        self.assertEqual([
-            ['run', 'company', ContactsFilter(), ['zzz']]
-            ], self.ds.calls)
-
-    def test_empty_filter(self):
-        """Test that ContactsFilter has all proper defaults."""
-        result = self.driver.build_filter("{}")
-        self.assertEquals(ContactsFilter(), result)
-
-    def test_city_filter(self):
-        """Test that city filter is built properly."""
-        result = self.driver.build_filter(
-                '{"locations": {"city": "Indy"}}')
-        self.assertEquals(ContactsFilter(locations={'city': 'Indy'}), result)
-
-    def test_date_filters(self):
-        """Test that date filters are built properly."""
-        spec = '{"date": ["2015-09-01", "2015-09-30"]}'
-        result = self.driver.build_filter(spec)
-        expected = ContactsFilter(
-            date=[datetime(2015, 9, 1), datetime(2015, 9, 30)])
-        self.assertEquals(expected, result)
-
-    def test_order(self):
-        """Test that order is built properly."""
-        self.assertEqual(['name'], self.driver.build_order('["name"]'))
-
-    def test_encode_filter_interface(self):
-        """Test that filter interface is serialized properly."""
-        report_config = ReportConfiguration([
-            ColumnConfiguration(
-                column='name',
-                format='text'),
-            ColumnConfiguration(
-                column='date',
-                format='us_date',
-                filter_interface='date_range',
-                filter_display='Date'),
-            ColumnConfiguration(
-                column='locations',
-                format='city_state_list',
-                filter_interface='city_state',
-                filter_display='Locations',
-                help=True),
-            ColumnConfiguration(
-                column='tags',
-                format='comma_sep',
-                filter_interface='search_multiselect',
-                filter_display='Tags',
-                help=True),
-            ColumnConfiguration(
-                column='partner',
-                format='text',
-                filter_interface='search_multiselect',
-                filter_display='Partners',
-                help=True),
-        ])
-
-        self.assertEquals({
-            'filters': [
-                {
-                    'display': 'Date',
-                    'filter': 'date',
-                    'interface_type': 'date_range',
-                },
-                {
-                    'display': 'Locations',
-                    'filter': 'locations',
-                    'interface_type': 'city_state',
-                },
-                {
-                    'display': 'Tags',
-                    'filter': 'tags',
-                    'interface_type': 'search_multiselect',
-                },
-                {
-                    'display': 'Partners',
-                    'filter': 'partner',
-                    'interface_type': 'search_multiselect',
-                },
-            ],
-            'help': {
-                'partner': True,
-                'tags': True,
-                'locations': True,
-            },
-        }, self.driver.encode_filter_interface(report_config))
-
-
-class TestContactsFilterCloning(TestCase):
-    def test_clone_without_empty(self):
-        """Test clearing filter fields on an empty filter."""
-        filter = ContactsFilter()
-        self.assertEqual(ContactsFilter(), filter.clone_without_city())
-        self.assertEqual(ContactsFilter(), filter.clone_without_state())
-
-    def test_clone_without_full(self):
-        """Test clearing filter fields that are populated."""
-        filter = ContactsFilter(
-                tags=['C'],
-                locations={'city': 'A', 'state': 'B'})
-        expected_with_city = ContactsFilter(
-                tags=['C'],
-                locations={'city': 'A'})
-        expected_with_state = ContactsFilter(
-                tags=['C'],
-                locations={'state': 'B'})
-        self.assertEqual(expected_with_state, filter.clone_without_city())
-        self.assertEqual(expected_with_city, filter.clone_without_state())
